Dan's refactor
