--- ftdc/sys/sys_common.go
@@ -1,41 +1,0 @@
-// Package sys provides functionality for gathering system metrics in an FTDC compliant API.
-package sys
-
-// Common types and interfaces shared between Unix and Windows implementations
-
-type stats struct {
-	UserCPUSecs     float64
-	SystemCPUSecs   float64
-	ElapsedTimeSecs float64
-	VssMB           float64
-	RssMB           float64
-}
-
-type netDevLine struct {
-	RxBytes   uint64
-	RxPackets uint64
-	RxErrors  uint64
-	RxDropped uint64
-	TxBytes   uint64
-	TxPackets uint64
-	TxErrors  uint64
-	TxDropped uint64
-}
-
-type ifaceStats struct {
-	TxQueueLength uint64
-	RxQueueLength uint64
-	UsedSockets   uint64
-	Drops         uint64
-}
-
-type networkStats struct {
-	Ifaces map[string]netDevLine
-	TCP    ifaceStats
-	UDP    ifaceStats
-}
-
-// Statser interface for process usage statistics
-type Statser interface {
-	Stats() any
-}
